violet: Disable memcg kernel and socket accounting
Disable memcg kernel and socket accounting on Talos since
memory cgroups is used in this target as a means of user space task
grouping based on oom_score_adj, and controlling the userspace
memory consumed by tasks. Kernel and socket memory accounting is
not considered. Since memcg is created per process on this target
enabling kernel accounting results in the creation of numerous
kmem caches resulting in significant overhead.

Change-Id: Ic2315a92a2e78828a4150732d96002cf8daec52c
Signed-off-by: raghavt20 <raghavtandon2000@gmail.com>